Fix date in README.
This program was written in 1997 (which makes sense as that corresponds
with my aunt's passing) but I put it on the web originally in 2001 and
did a large cleanup at that point. So all the logs record 2001 as the
date, but that was just the cleanup, not the original program date.